+++ src/utils.h
@@ -32,11 +32,11 @@ public:
                 m_amount(amount){};
 
     T
-    operator()(T const newValue, double const weight = 1.0f)
+    operator()(T const newValue, double const weight = 1.0)
     {
-        const float factor = std::pow(m_amount, weight);
+        const double factor = std::pow(m_amount, weight);
 
-        m_filteredValue = factor * m_filteredValue + (1.0f - factor) * newValue;
+        m_filteredValue = factor * m_filteredValue + (1.0 - factor) * newValue;
         return m_filteredValue;
     }
